setup test socketio #2
