+++ status/__init__.py
@@ -76,7 +76,7 @@ class SetStatus:
         if self.files:
             for f, value in self.files.items():
 
-                if "xlsx" not in f:
+                if "xlsx" not in f or app.testing is True:
                     f = self.format_String(f)
 
                 filesav = os.path.join(path_pid, f)
@@ -100,7 +100,7 @@ class SetStatus:
                 ws: Worksheet = wb.active
                 rows = ws.max_row
 
-        elif data.get("typebot") == "pauta":
+        elif data.get("typebot") == "pauta":  # pragma: no cover
             data_inicio_formated = datetime.strptime(
                 data.get("data_inicio"), "%Y-%m-%d"
             )
@@ -110,7 +110,7 @@ class SetStatus:
             diff = data_fim_formated - data_inicio_formated
             rows = diff.days + 2
 
-        elif data.get("typebot") == "proc_parte":
+        elif data.get("typebot") == "proc_parte":  # pragma: no cover
 
             rows = len(list(data.get("varas"))) + 1
 
@@ -123,7 +123,7 @@ class SetStatus:
         max_length = name_column.type.length
         xlsx_ = str(data.get("xlsx", "Sem Arquivo"))
 
-        if len(data.get("xlsx", "Sem Arquivo")) > int(max_length):
+        if len(data.get("xlsx", "Sem Arquivo")) > int(max_length):  # pragma: no cover
             xlsx_ = xlsx_[: int(max_length)]
 
         execut = Executions(
